requirements.txt
django-oscar
pysolr
django-compressor

